+++ ecs.php
@@ -3,10 +3,20 @@
 declare(strict_types=1);
 
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symplify\EasyCodingStandard\ValueObject\Option;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $parameters = $containerConfigurator->parameters();
 
-    $containerConfigurator->import(__DIR__ . '/vendor/lmc/coding-standard/ecs.php');
-    $containerConfigurator->import(__DIR__ . '/vendor/lmc/coding-standard/ecs-7.4.php');
+    $parameters->set(
+        Option::SKIP,
+        [
+            'PHP_CodeSniffer\Standards\Generic\Sniffs\PHP\ForbiddenFunctionsSniff.Found' => [
+                'src/Services/AppStatusCollector.php',
+            ],
+        ]
+    );
+
+    $containerConfigurator->import(__DIR__ . '/tools/coding-standards/vendor/lmc/coding-standard/ecs.php');
+    $containerConfigurator->import(__DIR__ . '/tools/coding-standards/vendor/lmc/coding-standard/ecs-7.4.php');
 };