+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 
 Unreleased
+- Add `env_file` option
+- Show environment details in profiler
+- Fix collector template
 
 ## 6.0.0 - 2022-07-15
 - Make private library public